requirements.txt
asgiref==3.4.1
blinker==1.4
can==0.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
data==0.4
decorator==5.0.9
Deprecated==1.2.12
Django==3.2.5
Flask==2.0.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
funcsigs==1.0.2
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonconfig-tool==1.5
keyring==23.0.1
MarkupSafe==2.0.1
messages==0.5.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
protobuf==3.17.3
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-box==5.3.0
python-dotenv==0.18.0
python-util==1.2.1
python-utils==2.5.6
pytz==2021.1
pywin32-ctypes==0.2.0
queries==2.1.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
uuid==1.30
validus==0.3.0
vonage==2.5.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
